Add output flag to both yaml and json formats.
